backend/requirements.txt
djangorestframework==3.16.1
pillow==11.3.0
djangorestframework-simplejwt==5.5.1
pytest-django==4.11.1